+++ Src/Common/Quaternion.h
@@ -9,27 +9,27 @@ public:
 	
 	static constexpr float kEpsilonNormalSqrt = 1e-15F;
 
-	double w;
-	double x;
-	double y;
-	double z;
+	float w;
+	float x;
+	float y;
+	float z;
 
 	Quaternion(void);
 	Quaternion(const VECTOR& rad);
-	Quaternion(double w, double x, double y, double z);
+	Quaternion(float w, float x, float y, float z);
 
 	~Quaternion(void);
 
 	// オイラー角からクォータニオンへ変換
 	static Quaternion Euler(const VECTOR& rad);
-	static Quaternion Euler(double radX, double radY, double radZ);
+	static Quaternion Euler(float radX, float radY, float radZ);
 
 	// クォータニオンの合成
 	static Quaternion Mult(const Quaternion& q1, const Quaternion& q2);
 	Quaternion Mult(const Quaternion& q) const;
 
 	// 指定軸を指定角分、回転させる
-	static Quaternion AngleAxis(double rad, VECTOR axis);
+	static Quaternion AngleAxis(float rad, VECTOR axis);
 
 	// 座標を回転させる
 	static VECTOR PosAxis(const Quaternion& q, VECTOR pos);
@@ -59,8 +59,8 @@ public:
 	VECTOR GetDown(void) const;
 
 	// 内積
-	static double Dot(const Quaternion& q1, const Quaternion& q2);
-	double Dot(const Quaternion& b) const;
+	static float Dot(const Quaternion& q1, const Quaternion& q2);
+	float Dot(const Quaternion& b) const;
 
 	// 正規化
 	static Quaternion Normalize(const Quaternion& q);
@@ -71,17 +71,17 @@ public:
 	Quaternion Inverse(void) const;
 
 	// 球面補間
-	static Quaternion Slerp(Quaternion from, Quaternion to, double t);
+	static Quaternion Slerp(Quaternion from, Quaternion to, float t);
 
 	// ２つのベクトル間の回転量を取得する
 	static Quaternion FromToRotation(VECTOR fromDir, VECTOR toDir);
 	static Quaternion RotateTowards(const Quaternion& from, const Quaternion& to, float maxDegreesDelta);
-	static double Angle(const Quaternion& q1, const Quaternion& q2);
+	static float Angle(const Quaternion& q1, const Quaternion& q2);
 	static Quaternion SlerpUnclamped(Quaternion a, Quaternion b, float t);
 	static Quaternion Identity(void);
 
-	double Length(void) const;
-	double LengthSquared(void) const;
+	float Length(void) const;
+	float LengthSquared(void) const;
 	VECTOR xyz(void) const;
 
 	// 対象方向の回転
